// sourcecode/BTrak.Packages/snova-payroll/module/module-components/src/lib/Payroll/models/company-registration-model.ts
export class CompanyRegistrationModel {
    countryId: string;
    countryName: string;
    currencyId: string;
    currencyName: string;
    numberFormatId: string;
    numberFormat: string;
    dateFormatId: string;
    dateFormatName: string;
    timeFormatId: string;
    timeFormatName: string;
    isArchived: boolean;
    timeStamp: any;
    companyName: string;
    workEmail: string;
    password: string;
    siteAddress: string;
    industryId: string;
    isRemoteAccess:boolean;
    mainUseCaseId: string;
    role: string;
    teamSize: number;
    mobileNumber: string;
    timeZoneId: string;
    firstName: string;
    lastName: string;
    isDemoData: boolean;
    isSoftWare: boolean;
}